requirements.txt
requests
flask
beautifulsoup4
python-whois
dnspython